Changed name of method from ConvertToFeature to ToFeature.

+++ Geode.Tests/Models/Flight.cs
@@ -9,6 +9,6 @@ namespace Geode.Tests.Models
         public int Id { get; set; }
         public string Name { get; set; }
         [Geometry(GeoType.LineString)]
-        public IEnumerable<IEnumerable<int>> FlightPath { get; set; }
+        public IEnumerable<IEnumerable<double>> FlightPath { get; set; }
     }
 }
